requirements.txt
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scipy
scikit-learn==1.6.1
imblearn
xgboost
catboost
pymongo
from_root
evidently==0.2.8
dill
PyYAML
neuro_mf
boto3
mypy-boto3-s3
botocore
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
mlflow
dagshub
